Put template plugin in separate file, and rename it. Add more tests
