+++ package.json
@@ -45,7 +45,7 @@
     "istanbul": "^0.4.3",
     "mockery": "^1.7.0",
     "remap-istanbul": "^0.6.4",
-    "sinon": "^1.17.5",
+    "sinon": "^2.0.0",
     "yargs": "^5.0.0"
   },
   "dependencies": {
